requirements.txt
streamlit
pandas
openpyxl
requests
beautifulsoup4
python-docx
jinja2
